+++ plugin_info/install.php
@@ -33,6 +33,7 @@ require_once dirname(__FILE__) . '/../../../core/php/core.inc.php';
     }	
 	
 	chmod("/var/www/html/plugins/teleo/resources/get_veolia_data.sh", 0755);
+	chmod("/var/www/html/plugins/teleo/resources/install_apt_deb10.sh", 0755);  
   }
 
 // Fonction exécutée automatiquement après la mise à jour du plugin
@@ -50,6 +51,7 @@ require_once dirname(__FILE__) . '/../../../core/php/core.inc.php';
     }
 
 	chmod("/var/www/html/plugins/teleo/resources/get_veolia_data.sh", 0755);
+	chmod("/var/www/html/plugins/teleo/resources/install_apt_deb10.sh", 0755); 
 	
 	$file = '/var/www/html/plugins/teleo/plugin_info/packages.json';
 	if (file_exists($file)) {